move dags e2e tests
