+++ .github/workflows/ci.yml
@@ -46,6 +46,8 @@ jobs:
         run: poetry run pytest integration_tests
       - name: Run data tests
         run: poetry run pytest dags/tests
+      - name: Run data e2e tests
+        run: poetry run pytest dags/e2e_tests
       - name: Check code formatting
         run: poetry run black --check --diff .
       - uses: astral-sh/ruff-action@eaf0ecdd668ceea36159ff9d91882c9795d89b49 # v3.4.0
