+++ dags/e2e_tests/test_e2e_dag_template_read_db.py
@@ -6,8 +6,8 @@ import pytest
 from airflow import DAG
 from airflow.operators.python import PythonOperator
 
+from dags.e2e_tests.e2e_utils import airflow_init, ti_get
 from dags.shared.config.start_dates import START_DATES
-from dags.tests.e2e.e2e_utils import airflow_init, ti_get
 
 airflow_init()
